refactor: UI 모델의 패키지 분리
