+++ app/src/main/java/woowacourse/shopping/uimodel/ProductListMapper.kt
@@ -1,6 +1,7 @@
-package woowacourse.shopping.productlist
+package woowacourse.shopping.uimodel
 
 import woowacourse.shopping.domain.Product
+import woowacourse.shopping.uimodel.ProductUiModel
 
 fun Product.toProductUiModel(): ProductUiModel =
     ProductUiModel(
