+++ app/src/main/res/layout/item_cart.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="cartItemUiModel"
-            type="woowacourse.shopping.shoppingcart.CartItemUiModel" />
+            type="woowacourse.shopping.uimodel.CartItemUiModel" />
 
         <variable
             name="clickListener"
